counters/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
hazelcast-python-client==5.4.0